pyproject.toml
[project]
name = "collab-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi-sse>=1.1.1",
    "psycopg>=3.3.2",
    "pydantic>=2.12.5",
    "redis>=7.1.0",
    "sqlmodel>=0.0.31",
    "sse-starlette>=3.2.0",
]  # optional, uv ignores this

[tool.uv]
# optional uv config goes here, leave empty if you don't need anything
